Deletion, rename and reorganize the internal packages
* Deleted the `OSUtils.java` (this class are no longer used)
* Moved all internal classes to directory (package) called `internal`
* Renamed the `Options` class to `JMatrixUtils`, improving clarity
